Numerology game
You need to create a program that psychologically determines what a person's life path is through numerology.

Follow the instructions below:
Create a class named Person

Give this class 5 attributes:
    1. first_name(str): Person's first name
    2. last_name(str): Person's last name
    3. age(int): The person's age
    4. email(str): Person's email
    5. birth_day: The person's birthday # is an attribute property that returns 3 integers

The following functions are implemented in the Person class:
Create a constructor to set all the initial values ​of the object
Create a method called get_info() that returns all of the person's information
Create a method called get_life_path_number() that returns the number generated by the person's date of birth.
How is this number (life_path_number) calculated?
For example, we have a person born on July 14, 2002, whose name is Jamshid.
Step 1: Day: 14 → 1 + 4 = 5
Step 2: Month: 07 → 0 + 7 = 7
Step 3: Year: 2002 → 2+0+0+2 = 4
Step 4: Sum: 5 + 7 + 4 = 16
Step 5: Sum: 16 → 1+6 = 7
Result: 7 ✅
Answer: Jamshid's life_path_number is 7.
           👆🏻 this algorithm should work dynamically. should not work only for the specified case !!!
Create a method that receives a message that corresponds to a number.
called get_info_by_number() and accepts a parameter named number.
You need to read the data corresponding to the number from the file.  Download the file
A # sign is placed in front of the number for ease of reading.

Additionally, the following should be done:
Ask the user for information (first_name, last_name, ...) through the console after the program is launched.
Create an object of the Person class and give it the requested information.
Call the get_info() method and display information about the person.
call the get_life_path_number() method and save the calculated number to a variable and output it to the console
Call the get_info_by_number() method and return the psychological advice (definition in the file) given to the user by the output number.